Move doc comments to header files Fixes #20 Fixes #13
