+++ src/PinManager.cpp
@@ -15,25 +15,21 @@ PinManager::PinManager(uint8_t numberOfPins)
   _registeredPins = new map<uint8_t, MFModuleType>;
 }
 
-// Sets all pins to kNotSet
 void PinManager::ClearRegisteredPins()
 {
   _registeredPins->clear();
 }
 
-// Returns true if the specified pin is registered
 bool PinManager::IsPinRegistered(ARDUINO_PIN arduinoPinName)
 {
   return _registeredPins->count(arduinoPinName) > 0;
 }
 
-// Registers a pin as the specified type
 void PinManager::RegisterPin(ARDUINO_PIN arduinoPinName, MFModuleType type)
 {
   (*_registeredPins)[arduinoPinName] = type;
 }
 
-// Returns the equivalent STM32 pin for a given Arduino pin
 std::optional<PinName> PinManager::MapArudinoPin(ARDUINO_PIN arduinoPin)
 {
   const auto stm32pin = PinMappings.find(arduinoPin);
